+++ sdk/src/com/appnexus/opensdk/BannerAdView.java
@@ -26,7 +26,6 @@ import android.content.res.TypedArray;
 import android.graphics.Point;
 import android.os.Build;
 import android.util.AttributeSet;
-import android.util.Pair;
 import android.view.Display;
 import android.view.Gravity;
 import android.view.View;
@@ -35,7 +34,6 @@ import android.view.WindowManager;
 import android.view.animation.Animation;
 import android.webkit.WebView;
 import android.widget.FrameLayout;
-import android.widget.ViewAnimator;
 
 import com.appnexus.opensdk.transitionanimation.Animator;
 import com.appnexus.opensdk.transitionanimation.TransitionDirection;
@@ -45,8 +43,6 @@ import com.appnexus.opensdk.utils.Settings;
 import com.appnexus.opensdk.utils.WebviewUtil;
 
 import java.lang.ref.WeakReference;
-import java.util.LinkedList;
-import java.util.Queue;
 
 
 /**
@@ -329,8 +325,9 @@ public class BannerAdView extends AdView {
     private Displayable currentDisplayable;
     @Override
     void interacted(){
-        return;
+        // do nothing
     }
+
 
     @Override
     protected void display(Displayable d) {