+++ sdk/src/com/appnexus/opensdk/MRAIDAdActivity.java
@@ -75,6 +75,11 @@ class MRAIDAdActivity implements AdActivity.AdActivityImplementation {
     }
 
     @Override
+    public void browserLaunched() {
+        // do nothing
+    }
+
+    @Override
     public WebView getWebView() {
         return webView;
     }